:white_check_mark: refatorando algumas funções
home
